server/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docopt==0.6.2
idna==3.3
numpy==1.21.3
Pillow==8.4.0
pipreqs==0.4.11
psycopg2-binary==2.9.1
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
urllib3==1.26.7
wincertstore==0.2
yarg==0.1.9
gunicorn